requirements.txt
aiogram==2.25.1
aiosignal==1.3.1
alembic==1.12.0
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.28.0
attrs==23.1.0
celery==5.2.7
cffi==1.16.0
Django==4.2.4
django-celery-results==2.5.1
django-common==0.1.51
django-crispy-forms==2.1
django-cron==0.6.0
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-mathfilters==1.0.0
django-phonenumber-field==7.2.0
django-redis==5.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
email-validator==2.0.0.post2
emoji==1.7.0
exceptiongroup==1.1.3
fake-useragent==1.0.1
flower==2.0.1
frozenlist==1.4.0
google-api-core==2.8.1
google-api-python-client==2.50.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
greenlet==3.0.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
Jinja2==3.1.2
kombu==5.3.3
lxml==5.1.0
magic-filter==1.0.11
Mako==1.2.4
MarkupSafe==2.1.3
oauthlib==3.2.2
opencv-python==4.9.0.80
orjson==3.9.9
packaging==23.2
phonenumbers==8.13.25
Pillow==10.0.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==22.3.5
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyreadline3==3.4.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
Python-IO==0.3
python-multipart==0.0.6
python-social-auth==0.3.6
python-telegram-bot==13.12
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
telebot==0.0.5
telegram==0.0.1
TgCrypto==1.2.5
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.23.2
wcwidth==0.2.10
websocket-client==1.6.4
websockets==11.0.3
yarl==1.9.2
django-environ

